Separate into two classes
