+++ atsd_webtests/src/test/java/com/axibase/webtest/CommonAssertions.java
@@ -4,8 +4,7 @@ import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 public class CommonAssertions {
 
@@ -17,5 +16,9 @@ public class CommonAssertions {
     public static void assertInvalid(String errorMessage, WebDriver driver, WebElement element) {
         assertFalse(errorMessage, (Boolean) ((JavascriptExecutor) driver).
                 executeScript("return arguments[0].checkValidity()", element));
+    }
+
+    public static void assertValueAttributeOfElement(String errorMessage, String correctValue, WebElement element) {
+        assertEquals(errorMessage, correctValue, element.getAttribute("value"));
     }
 }
